requirements.txt
google-cloud-pubsub==2.7.1
google-cloud-aiplatform==1.3.0
google-cloud-vision==2.4.2
google-cloud-language==2.2.2
google-cloud-storage==1.42.0
google-cloud-bigquery==2.26.0
google-cloud-datastore==2.1.6
Flask==2.0.1
gunicorn==20.1.0
googlemaps==4.5.3
